ad design and payment

+++ PersonalAudioAssistant.Contracts/AppSettings/AppSettingsResponse.cs
@@ -4,7 +4,7 @@
     {
         public required string? Id { get; set; }
         public required string Theme { get; set; }
-        public required int Balance { get; set; }
+        public required decimal Balance { get; set; }
         public required string? UserId { get; set; }
     }
 }
