+++ PersonalAudioAssistant.Domain/Entities/AppSettings.cs
@@ -5,7 +5,7 @@ namespace PersonalAudioAssistant.Domain.Entities
     public class AppSettings: BaseEntity
     {
         public required string Theme { get; set; }
-        public required int Balance { get; set; }
+        public required decimal Balance { get; set; }
         public required string? UserId { get; set; }
     }
 }
